Setup xloop device with XLOOP_CONFIGURE ioctl call
This feature allows the userspace tool xlosetup to completely setup a
loop device with a single ioctl call, removing the in-between state
where the device can be partially configured, eg. the loop device has
a backing file associated with it, but is reading from the wrong offset.

Besides removing the intermediate state, another big benefit of this
ioctl is that XLOOP_SET_STATUS can be slow. The main reason for this
slowness is that XLOOP_SET_STATUS(64) calls blk_mq_freeze_queue() to
freeze the associated queue. This requires waiting for RCU.