qcow2: Add lock for lookup cache and decompression
Race conditions can occur if reads are submitted from different
cgroups, as each cgroup has its own workqueue. Linearize the
according parts of the code using a mutex, as this will in the
worst case lead to performance on par with pre-5.15 versions
of xloop, which used one workqueue for all cgroups.