[KERNEL] Fix build on CentOS 9.5-but-newer
So far, breaking changes were accompanied with a bump of the minor
version, but not this time. Bewetween -432 and-435, the signature of
disk_force_media_change() changed (haha), but the version number stayed
the same.

Pick the completely arbitrary macro QUEUE_FLAG_SYNCHRONOUS which only
exists in -435 to tell the versions apart.